+++ activerecord/test/cases/finder_test.rb
@@ -265,6 +265,12 @@ class FinderTest < ActiveRecord::TestCase
     assert_equal [Account], accounts.collect(&:class).uniq
   end
 
+  def test_find_by_association_subquery
+    author = authors(:david)
+    assert_equal author.post, Post.find_by(author: Author.where(id: author))
+    assert_equal author.post, Post.find_by(author_id: Author.where(id: author))
+  end
+
   def test_take
     assert_equal topics(:first), Topic.take
   end